Added and refactored existing helper funcs moved extractAttribute into entity helper module, added 2 new helper funcs for displaying repeated attributes
